create a singleton patchmatch object for use in inpainting
This creates a thin wrapper to patchmatch which loads the module
on demand, respecting the global "trypatchmatch" option.
